# src/infra/database/models/User.js
'use strict';

module.exports = function (mongoose) {
  const { Schema } = mongoose;
  const UserSchema = new Schema({
    name: {
      type: String,
      required: true,
    },
    email: {
      type: String,
      required: true,
      unique: true,
    },
    phone: {
      type: String,
      unique: true,
    },
    password: {
      type: String,
      required: true,
    },
    createdAt: {
      type: Date,
      default: Date.now,
    },
  });
  const User = mongoose.model('user', UserSchema);

  return User;
};